refactor(patcher): move validation into seperate file
