+++ src/patcher/patcher.ts
@@ -72,14 +72,3 @@ export async function executeScripts(
     });
   }
 }
-
-export async function validateTargetDirectory(
-  directory: string
-): Promise<boolean> {
-  try {
-    const stats = await stat(join(directory, 'package.json'));
-    return true;
-  } catch (error) {
-    return false;
-  }
-}
